backend/requirements.txt
Django>=4.0
djangorestframework>=3.14
django-cors-headers>=4.3.1